feat: add support for certified time endpoint (#109)
This PR adds support for the `set_certified_time` endpoint of the
PocketIC Server. This endpoint allows developers to set the certified
time of the subnet and have it immediately reflected in query calls and
read state requests without needing to explicitly call the `tick`
method. Several new methods are exposed that leverage this endpoint:
`resetCertifiedTime`, `setCertifiedTime` and `advanceCertifiedTime`. All
of these methods correspond to the existing `resetTime`, `setTime` and
`advanceTime` methods.

Ref: #SDK-2001

---------

Co-authored-by: Kaia Peacock <kaia@peacock.dev>