ci: update pipelines (#113)
This PR fixes the release pipeline by moving the `npm` environment from
the `create_release_pr` pipeline to the `release` pipeline. The NPM
token contained in the `npm` environment is needed in the `release`
pipeline to publish to NPM.

The `release` pipeline was also missing steps to build the NPM package,
and generate release notes for the GitHub release artifact.